requirements.txt
asgiref==3.5.2
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
deckar01-ratelimit==3.0.2
deepl==1.11.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
docx==0.2.4
et-xmlfile==1.1.0
filelock==3.8.2
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
keybert==0.7.0
lxml==4.9.2
mysqlclient==2.1.1
nltk==3.8
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==9.3.0
Pygments==2.13.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
scikit-learn==1.2.0
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
types-pytz==2022.7.0.0
typing-extensions==4.4.0
urllib3==1.25.11
wordhoard==1.5.1
